python-apt: work around mangling of distributor ID in lsb-release
Due to Debian taking the liberty to mangle the distributor ID from
os-release in [1], and only attempting and even not releasing a
half-assed fix [2] to buster, we provide a "Bunsenlabs.info" distro
template file to work around some of the implications, which, among
other things, is apt-add-repository from software-properties-common
failing [3] as the output of lsb-release is used directly in python3-apt
to ID the current distro ID.

I'm too disappointed and too annoyed by bugs.debian.org to file a bug.
The bug -- accidentally -- affects neither Debian nor Ubuntu, because in
these special cases, the system works.

---

[1] https://salsa.debian.org/debian/lsb/-/blob/91f6838dd59660f1442c13f3df6260beeb78b148/lsb_release.py#L343
[2] https://salsa.debian.org/debian/lsb/-/commit/6225d576b2dd9d105d3ee48f053733c7ad072b5b
[3] https://forums.bunsenlabs.org/viewtopic.php?id=6925